Organize tests, add a couple more
