+++ cypress/support/index.js
@@ -18,3 +18,22 @@ import './commands';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
+
+const TEST_USERNAME = 'cypressTest';
+
+before(() => {
+	cy.visit('/').then(async () => {
+		cy.wait(500);
+		// Make sure user is logged in
+		cy.get('#welcomeMessage').then(element => {
+			if(element.is(':visible')) {
+				cy.get('#profilePanel').click();
+				cy.contains('Continue as guest').click();
+				cy.get('#usernamePickerText').type(TEST_USERNAME);
+				cy.contains('Confirm').click();
+			}
+		});
+	});
+});
+
+module.exports = { TEST_USERNAME };
